+++ src/pages/MainPage/MainPage.module.css
@@ -1,0 +1,161 @@
+/* 전체 컨테이너 스타일 */
+.container {
+  background-color: #0A0A0A;
+  color: #F5F5F5;
+  font-family: 'Noto Sans KR', sans-serif;
+  overflow-x: hidden;
+}
+
+/* 배경 효과 (파티클 등) */
+.backgroundEffect {
+  position: fixed;
+  top: 0; left: 0; width: 100vw; height: 100vh;
+  z-index: 0; pointer-events: none;
+  background: radial-gradient(ellipse at center, rgba(107, 33, 168, 0.15) 0%, transparent 60%);
+  opacity: 0.5;
+}
+
+/* 메인, 섹션 헤더 스타일 */
+main { position: relative; z-index: 1; }
+.sectionHeader { text-align: center; margin-bottom: 4rem; }
+.sectionHeader h2 { font-size: clamp(2.5rem, 5vw, 3.5rem); font-weight: 700; margin-bottom: 1rem; }
+.sectionHeader p { font-size: 1.2rem; color: #a0a0a0; max-width: 600px; margin: 0 auto; line-height: 1.7; }
+
+/* --- 1. 히어로(메인) 섹션 --- */
+.heroSection { height: 150vh; }
+.heroContent {
+  height: 100vh; width: 100%;
+  display: flex; flex-direction: column;
+  align-items: center; justify-content: center;
+  text-align: center; position: sticky; top: 0;
+  will-change: transform;
+}
+.heroContent h1 {
+  font-size: clamp(4rem, 12vw, 10rem);
+  font-weight: 800; letter-spacing: -5px; line-height: 1;
+  text-shadow: 0 0 30px rgba(255,255,255,0.2);
+}
+.heroContent p { font-size: 1.5rem; color: #888; margin-top: 2rem; max-width: 600px; }
+
+/* --- 2. 솔루션(문제/해결) 섹션 --- */
+.solutionSection {
+  min-height: 100vh; padding: 10rem 2rem;
+  display: flex; align-items: center; justify-content: center;
+  text-align: center; position: relative; z-index: 2;
+  background: #0A0A0A;
+}
+.solutionContent h2 {
+  font-size: clamp(3rem, 6vw, 5rem);
+  font-weight: 700; line-height: 1.3;
+}
+.highlight {
+  --highlight-width: 0%;
+  position: relative;
+  display: inline-block;
+  color: #F5F5F5;
+  transition: color 0.3s ease-in-out;
+}
+.highlight::after {
+  content: '';
+  position: absolute; bottom: -5px; left: 0;
+  width: var(--highlight-width);
+  height: 10px;
+  background: linear-gradient(90deg, #a855f7, #ec4899);
+}
+
+/* --- 3. How-To(사용법) 섹션 --- */
+.howToSection {
+  padding: 10rem 2rem;
+  position: relative; z-index: 2;
+  background-color: #111;
+  border-top: 1px solid #222; border-bottom: 1px solid #222;
+}
+.howToGrid {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
+  gap: 3rem; max-width: 1200px; margin: 0 auto;
+}
+.howToStep {
+  border: 1px solid #222;
+  padding: 2.5rem; border-radius: 16px;
+  background: #0A0A0A;
+}
+.stepNumber {
+  font-size: 2rem; font-weight: 700;
+  color: #a855f7; margin-bottom: 1rem;
+}
+.howToStep h3 { font-size: 1.8rem; font-weight: 600; margin-bottom: 1rem; }
+.howToStep p { font-size: 1rem; line-height: 1.7; color: #aaa; }
+
+/* --- 4. 가로 스크롤(워크플로우) 섹션 --- */
+.horizontalSection {
+  height: 100vh; width: 100%;
+  display: flex;
+}
+.featureCard {
+  width: 100vw; height: 100%;
+  flex-shrink: 0;
+  display: flex; justify-content: center; align-items: center;
+  padding: 5rem;
+  position: relative;
+}
+.cardContent {
+  max-width: 500px;
+  text-align: center;
+  padding: 3rem;
+  border-radius: 20px;
+  background: rgba(20, 20, 20, 0.5);
+  backdrop-filter: blur(15px);
+  -webkit-backdrop-filter: blur(15px);
+  border: 1px solid rgba(255, 255, 255, 0.1);
+  position: relative;
+}
+.cardContent::before { /* 오버레이로 가독성 향상 */
+  content: '';
+  position: absolute; top:0; left: 0; right: 0; bottom: 0;
+  border-radius: 20px;
+  background: linear-gradient(to top, rgba(0,0,0,0.5), transparent);
+  z-index: -1;
+}
+.cardContent svg { margin: 0 auto 2rem; color: #a855f7; }
+.cardContent h3 { font-size: 2.5rem; font-weight: 600; margin-bottom: 1rem; }
+.cardContent p { font-size: 1.1rem; line-height: 1.8; color: #ddd; }
+
+/* --- 5. Testimonial(사용자 피드백) 섹션 --- */
+.testimonialSection {
+  padding: 10rem 2rem;
+  background: #0A0A0A;
+}
+.testimonialGrid {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(400px, 1fr));
+  gap: 2rem; max-width: 1000px; margin: 0 auto;
+}
+.testimonialCard {
+  background-color: #111; border: 1px solid #222;
+  padding: 2.5rem; border-radius: 16px;
+  text-align: center;
+}
+.testimonialIcon { color: #ec4899; font-size: 2rem; margin-bottom: 1.5rem; }
+.testimonialCard p { font-size: 1.1rem; line-height: 1.8; font-style: italic; margin-bottom: 1.5rem; }
+.testimonialCard span { font-size: 1rem; font-weight: 600; color: #888; }
+
+/* --- 6. CTA(콜 투 액션) 섹션 --- */
+.ctaSection {
+  min-height: 90vh;
+  display: flex; flex-direction: column;
+  align-items: center; justify-content: center;
+  text-align: center; padding: 2rem;
+}
+.ctaSection h2 { font-size: clamp(3rem, 7vw, 5rem); font-weight: 700; line-height: 1.3; margin-bottom: 3rem; }
+.ctaButton {
+  display: inline-flex; align-items: center; gap: 0.75rem;
+  background: linear-gradient(90deg, #a855f7, #ec4899);
+  color: #fff; padding: 1.2rem 3rem; border-radius: 50px;
+  font-size: 1.2rem; font-weight: 600; text-decoration: none;
+  transition: all 0.3s ease;
+  box-shadow: 0 10px 30px -10px rgba(236, 72, 153, 0.5);
+}
+
+/* --- 7. 푸터 --- */
+.footer { text-align: center; padding: 4rem 2rem; color: #666; font-size: 0.9rem; }
